chore(refactor):
- rename convertObject.js to formatResponse.js
