+++ client/__tests__/store/MessageStore.spec.js
@@ -3,7 +3,7 @@ import 'babel-polyfill';
 import MessageStore from '../../src/stores/MessageStore';
 import getMessageResponse from '../../src/__mock__/getMessageResponse.json';
 import AppDispatcher from '../../src/dispatcher/AppDispatcher';
-import { helpGetGroupMessages } from '../../src/helper/convertObject';
+import { helpGetGroupMessages } from '../../src/helper/formatResponse';
 
 jest.mock('../../src/dispatcher/AppDispatcher');
 jest.dontMock('../../src/stores/MessageStore.js');
